Remove obsolete and overcomplicated code + unify tooltip color.
